requirements.txt
asgiref==3.8.1
click==8.1.8
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
uvicorn==0.34.0
whitenoise==6.9.0
